requirements.txt
python-dotenv
pymongo
pydantic
pandas
scikit-learn
numpy
scipy
nltk
spacy
langid
torch
transformers
sentence-transformers
fastapi
uvicorn
pyyaml
matplotlib
seaborn
selenium
webdriver-manager
textblob
gensim
sentencepiece
protobuf